+++ lib/AST/Types.cpp
@@ -289,19 +289,20 @@ const llvm::fltSemantics &FloatType::getAPFloatSemantics() const {
   llvm_unreachable("Unknown FloatKind!");
 }
 
-Optional<unsigned> TupleType::lookup(Identifier ident) const {
+Optional<size_t> TupleType::lookup(Identifier ident) const {
   IntegerWidth::Status status;
-  // Parse the identifier string as an arbitrary-width integer
+  // Parse the identifier string as an arbitrary-width integer written in base
+  // 10
   llvm::APInt value =
       IntegerWidth::arbitrary().parse(ident.c_str(), false, 10, &status);
   // If the value couldn't be parsed successfully, it can't be an integer.
   if (status != IntegerWidth::Status::Ok)
     return None;
   // The maximum index of the tuple is like an array: its size-1.
-  const unsigned numElem = getNumElements();
+  const size_t numElem = getNumElements();
   // If the value is greater or equal to that value, the index isn't legit,
   // else, return the parsed index.
-  unsigned result = value.getLimitedValue(numElem);
+  size_t result = value.getLimitedValue(numElem);
   if (result == numElem)
     return None;
   return result;
